Pointers and Arrays

+++ 0x05-pointers_arrays_strings/3-puts.c
@@ -7,11 +7,11 @@
  */
 void _puts(char *str)
 {
-	int i;
+	size_t i;
 
 	for (i = 0; i < strlen(str); i++)
 	{
 		_putchar(str[i]);
 	}
-	_putchar("\n");
+	_putchar('\n');
 }
